+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/ClientAuthRequest.kt
@@ -15,4 +15,8 @@ data class ClientAuthRequest(
     /** Validated [AuthorizationDetails] */
     val authnDetails: Collection<AuthorizationDetails>? = null,
     val codeChallenge: String? = null,
+    /** Client ID to which the auth code was issued */
+    val clientId: String? = null,
+    /** Authenticated with PAR */
+    val authenticatedClient: AuthenticatedClient? = null,
 )